requirements.txt
# =====================
# Core
# =====================
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0

# =====================
# Telegram Scraping
# =====================
telethon>=1.34
pytz>=2023.3

# =====================
# Database
# =====================
psycopg2-binary>=2.9
sqlalchemy>=2.0
alembic>=1.12

# =====================
# Data Transformation
# =====================
dbt-core>=1.6
dbt-postgres>=1.6

# =====================
# Computer Vision
# =====================
ultralytics>=8.0
opencv-python>=4.8
pillow>=10.0

# =====================
# API
# =====================
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.5

# =====================
# Orchestration
# =====================
dagster>=1.5
dagster-postgres>=0.20

# =====================
# Development
# =====================
jupyter>=1.0
black>=23.0
pytest>=7.4
pytest-cov>=4.1
